+++ packages/calcite-components/src/controllers/useFocusTrap.ts
@@ -28,7 +28,7 @@ export interface UseFocusTrap {
    *
    * @see https://github.com/focus-trap/focus-trap#trapupdatecontainerelements
    */
-  updateContainerElements: () => void;
+  updateContainerElements: (extraContainers?: ExtendedFocusTrapOptions["extraContainers"]) => void;
 }
 
 interface UseFocusTrapOptions<T extends LitElement = LitElement> {
@@ -43,6 +43,50 @@ interface UseFocusTrapOptions<T extends LitElement = LitElement> {
   focusTrapOptions?: FocusTrapOptions;
 }
 
+interface FocusTrapComponent extends LitElement {
+  /*
+   * When `true` prevents focus trapping.
+   */
+  focusTrapDisabled?: boolean;
+
+  /**
+   * When defined, provides a condition to disable focus trapping. When `true`, prevents focus trapping.
+   */
+  focusTrapDisabledOverride?: () => boolean;
+
+  /**
+   * Additional options to configure the focus trap.
+   */
+  focusTrapOptions?: ExtendedFocusTrapOptions;
+}
+
+export type ExtendedFocusTrapOptions =
+  /**
+   * @see https://github.com/focus-trap/focus-trap#createoptions
+   */
+  Pick<FocusTrapOptions, "allowOutsideClick" | "initialFocus" | "returnFocusOnDeactivate"> & {
+    /**
+     * Additional elements to include in the focus trap. This is useful for including elements that may have related parts rendered outside the main focus-trap element.
+     */
+    extraContainers: Parameters<FocusTrap["updateContainerElements"]>[0];
+  };
+
+function getEffectiveContainerElements(
+  targetEl: HTMLElement,
+  { focusTrapOptions }: FocusTrapComponent,
+  extraContainers?: ExtendedFocusTrapOptions["extraContainers"],
+) {
+  if (!focusTrapOptions?.extraContainers && !extraContainers) {
+    return targetEl;
+  }
+
+  return [targetEl, ...toContainerArray(focusTrapOptions?.extraContainers), ...toContainerArray(extraContainers)];
+}
+
+function toContainerArray(containers: ExtendedFocusTrapOptions["extraContainers"] = []) {
+  return Array.isArray(containers) ? containers : [containers];
+}
+
 /**
  * A controller for managing focus traps.
  *
@@ -50,19 +94,20 @@ interface UseFocusTrapOptions<T extends LitElement = LitElement> {
  *
  * @param options
  */
-export const useFocusTrap = <T extends LitElement>(
+export const useFocusTrap = <T extends FocusTrapComponent>(
   options: UseFocusTrapOptions<T>,
 ): ReturnType<typeof makeGenericController<UseFocusTrap, T>> => {
   return makeGenericController<UseFocusTrap, T>((component, controller) => {
     let focusTrap: FocusTrap;
     let focusTrapEl: HTMLElement;
-    const { focusTrapOptions } = options;
+    const internalFocusTrapOptions = options.focusTrapOptions;
 
     controller.onConnected(() => {
       if (component[options.triggerProp] && focusTrap) {
         focusTrap.activate();
       }
     });
+
     controller.onDisconnected(() => focusTrap?.deactivate());
 
     return {
@@ -74,10 +119,24 @@ export const useFocusTrap = <T extends LitElement>(
         }
 
         if (!focusTrap) {
-          focusTrap = createFocusTrap(targetEl, createFocusTrapOptions(targetEl, focusTrapOptions));
+          const effectiveFocusTrapOptions = {
+            ...internalFocusTrapOptions,
+            ...component.focusTrapOptions,
+          };
+
+          focusTrap = createFocusTrap(
+            getEffectiveContainerElements(targetEl, component),
+            createFocusTrapOptions(targetEl, effectiveFocusTrapOptions),
+          );
         }
 
-        focusTrap.activate(options);
+        if (
+          typeof component.focusTrapDisabledOverride === "function"
+            ? !component.focusTrapDisabledOverride()
+            : !component.focusTrapDisabled
+        ) {
+          focusTrap.activate(options);
+        }
       },
       deactivate: (options?: Parameters<FocusTrap["deactivate"]>[0]) => focusTrap?.deactivate(options),
       overrideFocusTrapEl: (el: HTMLElement) => {
@@ -87,9 +146,9 @@ export const useFocusTrap = <T extends LitElement>(
 
         focusTrapEl = el;
       },
-      updateContainerElements: () => {
+      updateContainerElements: (extraContainers?: ExtendedFocusTrapOptions["extraContainers"]) => {
         const targetEl = focusTrapEl || component.el;
-        return focusTrap?.updateContainerElements(targetEl);
+        return focusTrap?.updateContainerElements(getEffectiveContainerElements(targetEl, component, extraContainers));
       },
     };
   });